// Projecto actualizado_V2/backend/src/utilitarios/tipo-notificacao.enum.ts
export enum TipoNotificacao {
  LEMBRETE_DOACAO = 'DONATION_REMINDER',
  DOACAO_CONFIRMADA = 'DONATION_CONFIRMED',
  DOACAO_CANCELADA = 'DONATION_CANCELLED',
  PEDIDO_TRANSFUSAO = 'TRANSFUSION_REQUEST',
  TRANSFUSAO_CONCLUIDA = 'TRANSFUSION_COMPLETED',
  ANUNCIO_CAMPANHA = 'CAMPAIGN_ANNOUNCEMENT',
  URGENCIA_SANGUE = 'BLOOD_URGENCY',
  CONSULTA_AGENDADA = 'APPOINTMENT_SCHEDULED',
  RESULTADOS_TESTE = 'TEST_RESULTS_READY',
  ALERTA_SISTEMA = 'SYSTEM_ALERT',
  NOTIFICACAO_ADMIN = 'ADMIN_NOTIFICATION',
  TRIAGEM_EM_PROCESSAMENTO = 'TRIAGE_IN_PROCESSING',
  DOACAO_RECUSADA = 'DONATION_REFUSED',
  BOAS_VINDAS = 'WELCOME',
  MENSAGEM = 'MESSAGE',
}